+++ mobile/android/app/src/main/java/com/kostrzewa/cechini/model/ProductDTO.java
@@ -10,7 +10,7 @@ public class ProductDTO implements Serializable {
 
     private String name;
 
-    private Integer capacity;
+    private Double capacity;
 
     private String eanArt;
 
@@ -38,11 +38,11 @@ public class ProductDTO implements Serializable {
         this.name = name;
     }
 
-    public Integer getCapacity() {
+    public Double getCapacity() {
         return capacity;
     }
 
-    public void setCapacity(Integer capacity) {
+    public void setCapacity(Double capacity) {
         this.capacity = capacity;
     }
